fixed build, reinstated otel
