+++ apps/sim/.env.example
