+++ apps/sim/app/api/auth/oauth/connections/route.ts
@@ -40,7 +40,7 @@ export async function GET(request: NextRequest) {
       .where(eq(user.id, session.user.id))
       .limit(1)
 
-    const userEmail = userRecord.length > 0 ? userRecord[0].email : null
+    const userEmail = userRecord.length > 0 ? userRecord[0]?.email : null
 
     // Process accounts to determine connections
     const connections: any[] = []
