+++ apps/sim/app/w/[id]/components/control-bar/components/deploy-modal/components/chat-deploy/chat-deploy.tsx
@@ -31,6 +31,7 @@ import { Label } from '@/components/ui/label'
 import { Skeleton } from '@/components/ui/skeleton'
 import { Textarea } from '@/components/ui/textarea'
 import { env } from '@/lib/env'
+import { getNodeEnv } from '@/lib/environment'
 import { createLogger } from '@/lib/logs/console-logger'
 import { getBaseDomain } from '@/lib/urls/utils'
 import { cn } from '@/lib/utils'
@@ -54,7 +55,7 @@ interface ChatDeployProps {
 
 type AuthType = 'public' | 'password' | 'email'
 
-const isDevelopment = env.NODE_ENV === 'development'
+const isDevelopment = getNodeEnv() === 'development'
 
 const getDomainSuffix = (() => {
   const suffix = isDevelopment ? `.${getBaseDomain()}` : '.simstudio.ai'
