+++ apps/sim/lib/env.ts
@@ -1,8 +1,5 @@
 import { createEnv } from '@t3-oss/env-nextjs'
-import dotenv from 'dotenv'
 import { z } from 'zod'
-
-dotenv.config({ path: '../.env' })
 
 export const env = createEnv({
   server: {
