+++ apps/sim/package.json
@@ -9,20 +9,20 @@
     "node": ">=20.0.0"
   },
   "scripts": {
-    "dev": "dotenv -- next dev --turbo --port 3000",
-    "dev:classic": "dotenv -- next dev",
-    "build": "dotenv -- next build",
-    "start": "dotenv -- next start",
-    "lint": "dotenv -- next lint",
+    "dev": "next dev --turbo --port 3000",
+    "dev:classic": "next dev",
+    "build": "next build",
+    "start": "next start",
+    "lint": "next lint",
     "format": "prettier --write .",
     "format:check": "prettier --check .",
     "prepare": "cd ../.. && bun husky",
-    "db:push": "dotenv -- bunx drizzle-kit push",
-    "db:studio": "dotenv -- bunx drizzle-kit studio",
-    "test": "dotenv -- vitest run",
-    "test:watch": "dotenv -- vitest",
-    "test:coverage": "dotenv -- vitest run --coverage",
-    "email:dev": "dotenv -- email dev --dir components/emails",
+    "db:push": "bunx drizzle-kit push",
+    "db:studio": "bunx drizzle-kit studio",
+    "test": "vitest run",
+    "test:watch": "vitest",
+    "test:coverage": "vitest run --coverage",
+    "email:dev": "email dev --dir components/emails",
     "cli:build": "bun run build -w packages/simstudio",
     "cli:dev": "bun run build -w packages/simstudio && cd packages/simstudio && bun ./dist/index.js",
     "cli:publish": "cd packages/simstudio && bun publish",
