+++ package.json
@@ -5,7 +5,7 @@
     "dotenv-cli": "^8.0.0",
     "prettier": "^3.5.3",
     "prettier-plugin-tailwindcss": "^0.6.11",
-    "turbo": "^2.5.3"
+    "turbo": "2.5.3"
   },
   "packageManager": "bun@1.2.12",
   "name": "simstudio",
@@ -17,20 +17,20 @@
     "packages/*"
   ],
   "scripts": {
-    "build": "dotenv -- turbo run build",
-    "dev": "dotenv -- turbo run dev",
-    "test": "dotenv -- turbo run test",
+    "build": "turbo run build",
+    "dev": "turbo run dev",
+    "test": "turbo run test",
     "format": "prettier --write .",
     "format:check": "prettier --check .",
-    "lint": "dotenv -- turbo run lint"
+    "lint": "turbo run lint"
   },
   "overrides": {
     "react": "19.1.0",
-    "react-dom": "19.1.0"
+    "react-dom": "19.1.0",
+    "tailwindcss": "3.4.1"
   },
   "dependencies": {
     "@t3-oss/env-nextjs": "0.13.4",
-    "@types/bun": "^1.2.12",
-    "bun": "^1.2.13"
+    "@types/bun": "^1.2.12"
   }
 }
